+++ delphi-frontend/src/main/java/org/sonar/plugins/communitydelphi/api/ast/VarSectionNode.java
@@ -22,4 +22,6 @@ import java.util.List;
 
 public interface VarSectionNode extends DelphiNode {
   List<VarDeclarationNode> getDeclarations();
+
+  boolean isThreadVarSection();
 }